app/backend/requirements.txt
azure.identity
azure.keyvault.secrets
flask
kubernetes
requests